move render resolution function to lib
